backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
asgiref==3.11.1
async-timeout==5.0.1
attrs==26.1.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==48.0.0
Django==4.2.30
django-cors-headers==4.9.0
djangorestframework==3.16.1
frozenlist==1.8.0
http-ece==1.2.1
idna==3.13
multidict==6.7.1
mysqlclient==2.1.1
propcache==0.4.1
py-vapid==1.9.4
pycparser==2.23
pywebpush==2.1.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.22.0
